modularize l1_config
