+++ crates/madara/src/config/mod.rs
@@ -2,3 +2,5 @@ pub mod bootstrapper;
 pub mod constants;
 pub mod global_config;
 pub mod madara;
+
+mod l1_config;
